Made compatible with Python 3.
Added a default value to the sorted() call in process.py:38. Necessary because
a glyph with no 'pen down' ends up with a start of None, and comparing None to
a tuple during the sorted now causes a problem in python3.

Tested with python3.4 and 2.7, resulting in the same output file with both.